Set override values via JSON.
The previous implementation suffered from the following issues:
* `BOOL_KEY=False` would result in `BOOL_KEY` being set to `True`, because `bool('False') == True`
* Empty lists could not be set, as they were parsed as `['']`

With this CL, the following overrides become valid:
* `STRING_KEY="value"` - note that all strings have to be quoted
* `LIST_KEY=[false, 1, "string"]`
* `LIST_KEY=[]`
* `'DICT_KEY={"key1": 1.0, "key2": [16]}'` (the outer single quotes are necessary for Bash to treat the whole string as one argument).

We check the validity of JSON in the CLI client to fail fast.

Also, the code now raises an exception (rather than a warning) on missing keys or mismatching types, making errors more explicit.

PiperOrigin-RevId: 547787812
Change-Id: I508b9255f96fed8435fe1a5f5fb7c0748f3d1691