[Third_party] Rename `in_axis_resources` and `out_axis_resources` with `in_shardings` and `out_shardings`. This is just a simple name replacement. It does not change any of the current pjit semantics and doesn't break any code.
This is a safe and trivial name replacement. It does not change any of the semantics. You can still pass in PatitionSpecs to in_shardings and out_shardings.

The new names were introduced in jax 0.4.4, which was released Feb 16. If this code runs externally, you may need to bump your minimum jax version to 0.4.4.

PiperOrigin-RevId: 548267761
Change-Id: Ib9cc92b60c9d3e362feae5bc0772f5afa54acf8c