Pass `prng_key` as an argument to device computation
The current implementation embeds `prng_key` as a constant in the device computation, causing the device computation to change after every restart. This negatively affects the cacheability of the XLA compilation results as caching requires identical computation graphs.

This CL changes the device computation to take `prng_key` as an additional input so that the device computation remains the same even if `prng_key` changes across process restarts.

PiperOrigin-RevId: 602943226
Change-Id: I77076082912728fda66ceaf36155e149d833dc37