PR #25435: [export] Expand exporting to work with AbstractMesh.
Imported from GitHub PR https://github.com/jax-ml/jax/pull/25435

This is a follow up from #25640 that enables lowering with AbstractMesh.

This required adding `num_devices` to `lowering.compiler_args` because in presence of an AbstractMesh the device_assignment is not accurate.
Copybara import of the project:

--
afcb62ea20a2601ba1aa0cb844c9529c91d1980c by George Necula <gcnecula@gmail.com>:

[export] Expand exporting to work with AbstractMesh.

This is a follow up from #25640 that enabled lowering with
AbstractMesh.

This required adding `num_devices` to `lowering.compiler_args`
because in presence of an AbstractMesh the device_assignment
is not accurate.

Merging this change closes #25435

PiperOrigin-RevId: 706712699
Change-Id: I0ff762792205334e15e6361afc685fad890987a6